requirements.txt
streamlit==1.9.0
google_cloud_language==2.4.1
nltk==3.7
tweepy==4.9.0
spacy==3.3.0
scikit_learn==1.0.2
catboost==1.0.5